requirements.txt
# Core TPM library (choose one)
# Option 1: pytss (recommended for cross-platform)
pytss>=2.0.0
# Option 2: tpm2-pytss (alternative)
# tpm2-pytss>=2.0.0

# Cryptography
cryptography>=41.0.0

# Windows-specific (only needed on Windows)
pywin32>=305; sys_platform == 'win32'

# ASN.1 support (if needed)
pyasn1>=0.5.0
pyasn1-modules>=0.3.0

